conflict: store폴더 충돌 해결
